EZP-23966: add main search engine handler interface and move others down
