+++ eZ/Publish/SPI/Search/Content/Handler.php
@@ -1,0 +1,89 @@
+<?php
+/**
+ * File containing the Content Search handler interface
+ *
+ * @copyright Copyright (C) eZ Systems AS. All rights reserved.
+ * @license For full copyright and license information view LICENSE file distributed with this source code.
+ * @version //autogentag//
+ */
+
+namespace eZ\Publish\SPI\Search\Content;
+
+use eZ\Publish\SPI\Persistence\Content;
+use eZ\Publish\API\Repository\Values\Content\Query\Criterion;
+use eZ\Publish\API\Repository\Values\Content\Query;
+
+/**
+ * The Content Search handler retrieves sets of of Content objects, based on a
+ * set of criteria.
+ */
+interface Handler
+{
+    /**
+     * Finds content objects for the given query.
+     *
+     * @todo define structs for the field filters
+     *
+     * @throws \eZ\Publish\API\Repository\Exceptions\InvalidArgumentException if Query criterion is not applicable to its target
+     *
+     * @param \eZ\Publish\API\Repository\Values\Content\Query $query
+     * @param array $fieldFilters - a map of filters for the returned fields.
+     *        Currently supported: <code>array("languages" => array(<language1>,..))</code>.
+     *
+     * @return \eZ\Publish\API\Repository\Values\Content\Search\SearchResult With ContentInfo as SearchHit->valueObject
+     */
+    public function findContent( Query $query, array $fieldFilters = array() );
+
+    /**
+     * Performs a query for a single content object
+     *
+     * @throws \eZ\Publish\API\Repository\Exceptions\NotFoundException if the object was not found by the query or due to permissions
+     * @throws \eZ\Publish\API\Repository\Exceptions\InvalidArgumentException if Criterion is not applicable to its target
+     * @throws \eZ\Publish\API\Repository\Exceptions\InvalidArgumentException if there is more than than one result matching the criterions
+     *
+     * @todo define structs for the field filters
+     * @param \eZ\Publish\API\Repository\Values\Content\Query\Criterion $filter
+     * @param array $fieldFilters - a map of filters for the returned fields.
+     *        Currently supported: <code>array("languages" => array(<language1>,..))</code>.
+     *
+     * @return \eZ\Publish\SPI\Persistence\Content\ContentInfo
+     */
+    public function findSingle( Criterion $filter, array $fieldFilters = array() );
+
+    /**
+     * Suggests a list of values for the given prefix
+     *
+     * @param string $prefix
+     * @param string[] $fieldpath
+     * @param int $limit
+     * @param \eZ\Publish\API\Repository\Values\Content\Query\Criterion $filter
+     */
+    public function suggest( $prefix, $fieldPaths = array(), $limit = 10, Criterion $filter = null );
+
+    /**
+     * Indexes a content object
+     *
+     * @param \eZ\Publish\SPI\Persistence\Content $content
+     *
+     * @return void
+     */
+    public function indexContent( Content $content );
+
+    /**
+     * Deletes a content object from the index
+     *
+     * @param int $contentId
+     * @param int|null $versionId
+     *
+     * @return void
+     */
+    public function deleteContent( $contentId, $versionId = null );
+
+    /**
+     * Deletes a location from the index
+     *
+     * @param mixed $locationId
+     * @param mixed $contentId
+     */
+    public function deleteLocation( $locationId, $contentId );
+}
